Adds LineRenderers to Tile prefab
Adds two line renderers to the tile prefab -- one for the right grid
line, and one for the bottom grid line. This will be disabled depending
on the grid.

Currently, the prefab is called Tile2. That's because of my Unity
noob-ism. I want to get that back to being called Tile.