Adds Game Manager class
In theory, this is a singleton class which controls the overall game.
However, the singleton aspect doesn't seem to be working, as it's being
destroyed at the end of each scene.

Also, this check in fixes a couple of bugs:

1. On a replay, UpdateGridData was being called on the old grid. The
reason being that I hadn't removed the old function from the event
manager when the scene was over.

2. SetMaterial was crashing if OnTileOver was being called before the
activematerial had been set. The fix was to simply check for a null
material.