Ensures Game Manager persists between scenes
The Game Manager was being created as a component within the first scene
that called it. However, it was then being destroyed as soon as that
scene ended.

To fix this, I've simply called DoNotDestroyOnLoad in Awake, which
forces it to persist. That said, I'm not convinced that the class needs
to be a game component at all. We'll see if there are any benefits.