Fixes Stalemate Detection Bug
Fixes stalemate detection for cloned grid data. This wasn't working at
all, because I wasn't updating the move count when cloning the grid. As
a result, the stalemate detection -- which compared move count to number
of cells -- was never returning true.

This meant that the minimax algorithm wasn't working properly, as it
would never detect a 'draw' state.

To fix this, the stalemate detection no longer uses a move count.
Instead, it simply goes through the grid and -- if none are empty --
returns true.