Speeds up CPU moves
In particular the opening CPU move. This was partially being slowed down
by an artificial delay that I had added, which was to make the CPU move
seem more realistic. I've changed it so that this delay is only added if
necessary, i.e. if the CPU move was calculated within half a second.